+++ BostonData/Speech Assets/intent_schema.json
@@ -1,0 +1,40 @@
+{
+  "intents": [
+    {
+      "slots": [
+        {
+          "name": "Address",
+          "type": "AMAZON.PostalAddress"
+        }
+      ],
+      "intent": "SetAddressIntent"
+    },
+    {
+      "intent": "GetAddressIntent"
+    },
+    {
+      "intent": "TrashDayIntent"
+    },
+    {
+      "intent": "WorkZonesOnMyStreetIntent"
+    },
+    {
+      "intent": "WorkZonesOnAnyStreetIntent",
+      "slots": [
+        {
+          "name": "TargetAddress",
+          "type": "AMAZON.PostalAddress"
+        }
+      ]
+    },
+    {
+      "intent": "AMAZON.HelpIntent"
+    },
+    {
+      "intent": "AMAZON.StopIntent"
+    },
+    {
+      "intent": "AMAZON.CancelIntent"
+    }
+  ]
+}
