+++ BostonData/Speech Assets/sample_utterances.txt
@@ -1,0 +1,35 @@
+SetAddressIntent my address is {Address}
+SetAddressIntent i live at {Address}
+SetAddressIntent my home address is {Address}
+SetAddressIntent set address {Address}
+SetAddressIntent set my address to {Address}
+SetAddressIntent change address to {Address}
+SetAddressIntent change my address to {Address}
+SetAddressIntent update address to {Address}
+SetAddressIntent new address is {Address}
+GetAddressIntent what is my address set to
+GetAddressIntent what is my home address
+GetAddressIntent where do i live
+GetAddressIntent current address
+GetAddressIntent what is my current address
+GetAddressIntent give me my address
+TrashDayIntent when is my trash day
+TrashDayIntent when is garbage day
+TrashDayIntent what day is recycling
+TrashDayIntent recycling
+TrashDayIntent trash
+TrashDayIntent garbage
+TrashDayIntent what day should i put trash out
+TrashDayIntent which day is recycling
+TrashDayIntent when should i put out my trash
+TrashDayIntent when is garbage pickup
+TrashDayIntent garbage collection day
+WorkZonesOnMyStreetIntent is there any road work going on near me
+WorkZonesOnMyStreetIntent is there any street work going on near me
+WorkZonesOnMyStreetIntent is there any road work happening near me
+WorkZonesOnMyStreetIntent is there any street work happening near me
+WorkZonesOnAnyStreetIntent is there any road work happening on {TargetAddress}
+WorkZonesOnAnyStreetIntent is there any street work happening on {TargetAddress}
+WorkZonesOnAnyStreetIntent is there any road work going on {TargetAddress}
+WorkZonesOnAnyStreetIntent is there any street work going on {TargetAddress}
+WorkZonesOnAnyStreetIntent is there any work going on {TargetAddress}
